requirements.txt
asgiref==3.7.2
Deprecated==1.2.14
Django==3.2.15
django-db-connection-pool==1.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
greenlet==3.0.1
gunicorn==21.2.0
importlib-metadata==6.8.0
opentelemetry-api==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-wsgi==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparams==5.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
wrapt==1.16.0
zipp==3.17.0
